requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
_pytorch_select=0.2=gpu_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
blas=1.0=mkl
bottleneck=1.3.4=py37hce1f21e_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.7.22=hbcca054_0
certifi=2023.7.22=pyhd8ed1ab_0
cffi=1.15.0=py37hd667e15_1
chardet=3.0.4=py37h06a4308_1003
click=8.0.4=py37h06a4308_0
cryptography=36.0.0=py37h9ce1e76_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.10.0=py37_0
dataclasses=0.8=pyh6d0b6a4_7
dbus=1.13.18=hb2f20db_0
decorator=4.4.2=py_0
entrypoints=0.4=pyhd8ed1ab_0
expat=2.2.10=he6710b0_2
ffmpeg=4.3=hf484d3e_0
filelock=3.6.0=pyhd3eb1b0_0
fontconfig=2.13.0=h9420a91_0
freetype=2.10.4=h5ab3b9f_0
glib=2.66.1=h92f7085_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb31296c_0
huggingface_hub=0.4.0=py_0
icu=58.2=he6710b0_3
idna=2.10=pyhd3eb1b0_0
importlib-metadata=4.8.2=py37h06a4308_0
importlib_metadata=4.8.2=hd3eb1b0_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=5.5.5=py37h085eea5_0
ipython=7.33.0=py37h89c1867_0
ipython_genutils=0.2.0=py_1
jedi=0.19.1=pyhd8ed1ab_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9b=h024ee3a_2
jupyter_client=5.3.4=py37_0
jupyter_core=4.5.0=py_0
kiwisolver=1.3.0=py37h2531618_0
lame=3.100=h7b6447c_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.3.0=hdf63c60_0
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.19.1=h4ff587b_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.1.0=h2733197_1
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
lz4-c=1.9.2=heb0550a_3
matplotlib=3.3.2=0
matplotlib-base=3.3.2=py37h817c723_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mkl=2020.2=256
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.2.0=py37h23d657b_0
mkl_random=1.1.1=py37h0573a6f_0
munch=2.5.0=pyhd3eb1b0_0
ncurses=6.2=he6710b0_1
nettle=3.7.3=hbbd107a_1
networkx=2.5=py_0
ninja=1.10.1=py37hfd86e86_0
nltk=3.7=pyhd3eb1b0_0
numexpr=2.7.3=py37hb2eb853_0
numpy=1.19.2=py37h54aff64_0
numpy-base=1.19.2=py37hfa32c7d_0
olefile=0.46=py37_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1w=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.1=py37h8c16a72_0
parso=0.8.3=pyhd8ed1ab_0
pcre=8.44=he6710b0_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=8.0.1=py37he98fc37_0
pip=20.2.4=py37_0
prompt-toolkit=3.0.39=pyha770c72_0
protobuf=3.19.1=py37h295c915_0
ptyprocess=0.7.0=pyhd3deb0d_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.16.1=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py37h05f1152_2
pysocks=1.7.1=py37_1
python=3.7.9=h7579374_0
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytorch=1.11.0=py3.7_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pytz=2022.7=py37h06a4308_0
pyyaml=6.0=py37h7f8727e_1
pyzmq=19.0.2=py37hac76be4_2
qt=5.9.7=h5867ecd_1
readline=8.0=h7b6447c_0
regex=2021.11.2=py37h7f8727e_0
requests=2.23.0=py37_0
sacremoses=master=py_0
scikit-learn=0.23.2=py37h0573a6f_0
scipy=1.5.2=py37h0b6359f_0
setuptools=50.3.0=py37hb0f4dca_1
sip=4.19.8=py37hf484d3e_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.33.0=h62c20be_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hbc83047_0
tokenizers=0.10.3=py37_0
torchaudio=0.11.0=py37_cu113
torchvision=0.12.0=py37_cu113
tornado=6.0.4=py37h7b6447c_1
tqdm=4.62.3=pyhd3eb1b0_1
traitlets=5.9.0=pyhd8ed1ab_0
transformers=4.11.3=py_0
typing-extensions=3.10.0.2=hd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
urllib3=1.25.8=py37_0
wcwidth=0.2.8=pyhd8ed1ab_0
wheel=0.35.1=py_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zeromq=4.3.4=h2531618_0
zipp=3.7.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
